+++ src/features/posts/model/useSelectedPostStore.ts
@@ -1,5 +1,5 @@
 import { create } from 'zustand';
-import type { Post } from '../../types';
+import type { Post } from '../../../entities/posts/model';
 import { createStoreSelector } from '../../../shared/lib';
 
 // selectedPost 전역 상태 사용을 위한 hook
