+++ src/features/tags/api/useTagsQueries.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 
-import type { Tag } from '../../types';
+import type { Tag } from '../../../entities/tags/model/Tag';
 import { get } from '../../../shared/api/fetchBased';
 
 import { tagsQueryKeys } from '../config/tagQueryKeys';
